Modified the Endpoint objects to inherit from BaseAPI object and improved Exception handing in BaseAPI
